+++ src/cmd/mod.rs
@@ -10,3 +10,4 @@
 pub mod packages;
 pub mod path;
 pub mod sources;
+pub mod config;