+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [0.6.0] - 2018-03-01
+### Added
+- Plugin packages can now be declared in the `plugins` section of the config.
+- Add the `package` field to source groups.
+- Add the `bender config` command to dump the tool configuration.
+
 ## [0.5.0] - 2018-02-22
 ### Added
 - Add plugin support. The manifest may now have a `plugins` section.